Hoàn thành api trả lời câu hỏi từ người dùng

+++ src/main/java/studentConsulting/repository/RoleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import studentConsulting.model.entity.authentication.RoleEntity;
 
 @Repository
-public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
+public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
 
     RoleEntity findByName(String name);
 }
